2016-2017 Migration Start

+++ Leap/vtkLeapMarkerWidget.h
@@ -156,7 +156,7 @@ public:
 
   double global_CameraPosition;
 
-  int64_t global_ScaleFactorID ;  // Last  Frames Counted.
+  int global_ScaleFactorID ;  // Last  Frames Counted.
   double scaling_Max;
   double scaling_Start;
   double scaling_Min;
